requirements.txt
agno
requests
python-dotenv
pandas
psycopg2-binary
pgvector
beautifulsoup4
numpy
logging